Refax: moved buffer tools to a separate file
