+++ srtcore/buffer_snd.cpp
@@ -52,7 +52,6 @@ modified by
 
 #include "platform_sys.h"
 
-#include <cstring>
 #include <cmath>
 #include "buffer_snd.h"
 #include "packet.h"
@@ -64,103 +63,6 @@ namespace srt {
 using namespace std;
 using namespace srt_logging;
 using namespace sync;
-
-// You can change this value at build config by using "ENFORCE" options.
-#if !defined(SRT_MAVG_SAMPLING_RATE)
-#define SRT_MAVG_SAMPLING_RATE 40
-#endif
-
-bool AvgBufSize::isTimeToUpdate(const time_point& now) const
-{
-    const int      usMAvgBasePeriod = 1000000; // 1s in microseconds
-    const int      us2ms            = 1000;
-    const int      msMAvgPeriod     = (usMAvgBasePeriod / SRT_MAVG_SAMPLING_RATE) / us2ms;
-    const uint64_t elapsed_ms       = count_milliseconds(now - m_tsLastSamplingTime); // ms since last sampling
-    return (elapsed_ms >= msMAvgPeriod);
-}
-
-void AvgBufSize::update(const steady_clock::time_point& now, int pkts, int bytes, int timespan_ms)
-{
-    const uint64_t elapsed_ms       = count_milliseconds(now - m_tsLastSamplingTime); // ms since last sampling
-    m_tsLastSamplingTime            = now;
-    const uint64_t one_second_in_ms = 1000;
-    if (elapsed_ms > one_second_in_ms)
-    {
-        // No sampling in last 1 sec, initialize average
-        m_dCountMAvg      = pkts;
-        m_dBytesCountMAvg = bytes;
-        m_dTimespanMAvg   = timespan_ms;
-        return;
-    }
-
-    //
-    // weight last average value between -1 sec and last sampling time (LST)
-    // and new value between last sampling time and now
-    //                                      |elapsed_ms|
-    //   +----------------------------------+-------+
-    //  -1                                 LST      0(now)
-    //
-    m_dCountMAvg      = avg_iir_w<1000, double>(m_dCountMAvg, pkts, elapsed_ms);
-    m_dBytesCountMAvg = avg_iir_w<1000, double>(m_dBytesCountMAvg, bytes, elapsed_ms);
-    m_dTimespanMAvg   = avg_iir_w<1000, double>(m_dTimespanMAvg, timespan_ms, elapsed_ms);
-}
-
-int round_val(double val)
-{
-    return static_cast<int>(round(val));
-}
-
-CRateEstimator::CRateEstimator()
-    : m_iInRatePktsCount(0)
-    , m_iInRateBytesCount(0)
-    , m_InRatePeriod(INPUTRATE_FAST_START_US) // 0.5 sec (fast start)
-    , m_iInRateBps(INPUTRATE_INITIAL_BYTESPS)
-{}
-
-void CRateEstimator::setInputRateSmpPeriod(int period)
-{
-    m_InRatePeriod = (uint64_t)period; //(usec) 0=no input rate calculation
-}
-
-void CRateEstimator::updateInputRate(const time_point& time, int pkts, int bytes)
-{
-    // no input rate calculation
-    if (m_InRatePeriod == 0)
-        return;
-
-    if (is_zero(m_tsInRateStartTime))
-    {
-        m_tsInRateStartTime = time;
-        return;
-    }
-    else if (time < m_tsInRateStartTime)
-    {
-        // Old packets are being submitted for estimation, e.g. during the backup link activation.
-        return;
-    }
-
-    m_iInRatePktsCount  += pkts;
-    m_iInRateBytesCount += bytes;
-
-    // Trigger early update in fast start mode
-    const bool early_update = (m_InRatePeriod < INPUTRATE_RUNNING_US) && (m_iInRatePktsCount > INPUTRATE_MAX_PACKETS);
-
-    const uint64_t period_us = count_microseconds(time - m_tsInRateStartTime);
-    if (early_update || period_us > m_InRatePeriod)
-    {
-        // Required Byte/sec rate (payload + headers)
-        m_iInRateBytesCount += (m_iInRatePktsCount * CPacket::SRT_DATA_HDR_SIZE);
-        m_iInRateBps = (int)(((int64_t)m_iInRateBytesCount * 1000000) / period_us);
-        HLOGC(bslog.Debug,
-              log << "updateInputRate: pkts:" << m_iInRateBytesCount << " bytes:" << m_iInRatePktsCount
-                  << " rate=" << (m_iInRateBps * 8) / 1000 << "kbps interval=" << period_us);
-        m_iInRatePktsCount  = 0;
-        m_iInRateBytesCount = 0;
-        m_tsInRateStartTime = time;
-
-        setInputRateSmpPeriod(INPUTRATE_RUNNING_US);
-    }
-}
 
 CSndBuffer::CSndBuffer(int size, int maxpld, int authtag)
     : m_BufLock()
@@ -652,6 +554,13 @@ void CSndBuffer::ackData(int offset)
 int CSndBuffer::getCurrBufSize() const
 {
     return m_iCount;
+}
+
+namespace {
+int round_val(double val)
+{
+    return static_cast<int>(round(val));
+}
 }
 
 int CSndBuffer::getAvgBufSize(int& w_bytes, int& w_tsp)
