+++ srtcore/buffer_snd.h
@@ -50,14 +50,12 @@ modified by
    Haivision Systems Inc.
 *****************************************************************************/
 
-#ifndef INC_SRT_BUFFER_H
-#define INC_SRT_BUFFER_H
+#ifndef INC_SRT_BUFFER_SND_H
+#define INC_SRT_BUFFER_SND_H
 
-#include "udt.h"
-#include "list.h"
-#include "queue.h"
-#include "tsbpd_time.h"
-#include "utilities.h"
+#include "srt.h"
+#include "packet.h"
+#include "buffer_tools.h"
 
 // The notation used for "circular numbers" in comments:
 // The "cicrular numbers" are numbers that when increased up to the
@@ -72,76 +70,6 @@ modified by
 // a == b : equality is same as for just numbers
 
 namespace srt {
-
-/// The AvgBufSize class is used to calculate moving average of the buffer (RCV or SND)
-class AvgBufSize
-{
-    typedef sync::steady_clock::time_point time_point;
-
-public:
-    AvgBufSize()
-        : m_dBytesCountMAvg(0.0)
-        , m_dCountMAvg(0.0)
-        , m_dTimespanMAvg(0.0)
-    {
-    }
-
-public:
-    bool isTimeToUpdate(const time_point& now) const;
-    void update(const time_point& now, int pkts, int bytes, int timespan_ms);
-
-public:
-    inline double pkts() const { return m_dCountMAvg; }
-    inline double timespan_ms() const { return m_dTimespanMAvg; }
-    inline double bytes() const { return m_dBytesCountMAvg; }
-
-private:
-    time_point m_tsLastSamplingTime;
-    double     m_dBytesCountMAvg;
-    double     m_dCountMAvg;
-    double     m_dTimespanMAvg;
-};
-
-/// The class to estimate source bitrate based on samples submitted to the buffer.
-/// Is currently only used by the CSndBuffer.
-class CRateEstimator
-{
-    typedef sync::steady_clock::time_point time_point;
-    typedef sync::steady_clock::duration   duration;
-public:
-    CRateEstimator();
-
-public:
-    uint64_t getInRatePeriod() const { return m_InRatePeriod; }
-
-    /// Retrieve input bitrate in bytes per second
-    int getInputRate() const { return m_iInRateBps; }
-
-    void setInputRateSmpPeriod(int period);
-
-    /// Update input rate calculation.
-    /// @param [in] time   current time in microseconds
-    /// @param [in] pkts   number of packets newly added to the buffer
-    /// @param [in] bytes  number of payload bytes in those newly added packets
-    ///
-    /// @return Current size of the data in the sending list.
-    void updateInputRate(const time_point& time, int pkts = 0, int bytes = 0);
-
-    void resetInputRateSmpPeriod(bool disable = false) { setInputRateSmpPeriod(disable ? 0 : INPUTRATE_FAST_START_US); }
-
-private:                                                       // Constants
-    static const uint64_t INPUTRATE_FAST_START_US   = 500000;  //  500 ms
-    static const uint64_t INPUTRATE_RUNNING_US      = 1000000; // 1000 ms
-    static const int64_t  INPUTRATE_MAX_PACKETS     = 2000;    // ~ 21 Mbps of 1316 bytes payload
-    static const int      INPUTRATE_INITIAL_BYTESPS = BW_INFINITE;
-
-private:
-    int        m_iInRatePktsCount;  // number of payload bytes added since InRateStartTime
-    int        m_iInRateBytesCount; // number of payload bytes added since InRateStartTime
-    time_point m_tsInRateStartTime;
-    uint64_t   m_InRatePeriod;  // usec
-    int        m_iInRateBps;    // Input Rate in Bytes/sec
-};
 
 class CSndBuffer
 {
