+++ srtcore/filelist.maf
@@ -4,6 +4,7 @@ SOURCES
 api.cpp
 buffer_snd.cpp
 buffer_rcv.cpp
+buffer_tools.cpp
 cache.cpp
 channel.cpp
 common.cpp
@@ -55,6 +56,7 @@ PRIVATE HEADERS
 api.h
 buffer_snd.h
 buffer_rcv.h
+buffer_tools.h
 cache.h
 channel.h
 common.h
